requirements.txt
discord.py-self==2.0.0
imaplib2==3.6